# unicode/utf8: make DecodeRune{,InString} inlineable