This change makes the fast path for ASCII characters inlineable in
DecodeRune and DecodeRuneInString. Here are some benchmark results
(no change to allocations):

goos: darwin
goarch: amd64
pkg: unicode/utf8
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                             │     old      │                 new                  │
                             │    sec/op    │    sec/op     vs base                │
DecodeASCIIRune-8              2.4585n ± 1%   0.6187n ± 1%  -74.83% (p=0.000 n=20)
DecodeJapaneseRune-8            3.987n ± 1%    3.984n ± 1%        ~ (p=0.941 n=20)
DecodeASCIIRuneInString-8      2.2945n ± 1%   0.3123n ± 2%  -86.39% (p=0.000 n=20)
DecodeJapaneseRuneInString-8    3.821n ± 1%    3.983n ± 1%   +4.25% (p=0.000 n=20)
geomean                         3.045n         1.323n       -56.53%

Note: when #61502 gets resolved, this change should likely be reverted.

Updates #48195

---
🔄 **This is a mirror of upstream PR #75181**